Sessions and finished login and register forms
Added the session table to the database.
Added session creation and retrieving in the api and db modules.
Added the /api/session/ endpoint.
Logging in now sends a request for a new session, accepting a username and password.
Registering now also automatically logs in on top of creating new user data.
The forms' submit button now calls a function instead of sending a direct request.