Merged session into user table and added thread posting
Modified the GetUser API function to accept getting user information from the user's id, username, and session key.
Allow the frontend to access user information through querying the user, id, or if nothing at all, it will use the session key.
Added the post table that contains the posts that users will submit for discussion.
Deleted the session table and added the session key to the user table.
Set the primary key of the user table to be an auto incrementing id number instead of the username.
Added a submit page where users can submit posts.
Frontend now receives user information in the form of JSON.
Added the /api/threads/ endpoint for retrieving all the posts with no parent.