User page and post replies
Add user page that shows user information in kind of an ugly way.
Allowed multiline input and rendering.
Added a ReplyComponent that renders the reply form and post's children (recursively).
Add author links that navigates to the user's page.
Allow the user to click on a child's post and view it as the top PostComponent, and even reply to it as well.
Added the /api/children/ endpoint that returns a post's children, recursively.